[FIX] l10n_es_partner: Inject _rec_names_search properly
Previous patch was overriding _auto_init method for injecting the
comercial field in the list of fields to search for (_rec_names_search),
but this method is only called on module installation/update, so if
you restart the Odoo server after that, you lose the injection.

Thus, let's inject it if not present when calling name_search, and
this way we always have it available.

TT55241